@@ -2,6 +2,7 @@
   "name": "backend",
   "version": "0.0.0",
   "main": "server.js",
+  "type": "module",
   "private": true,
   "scripts": {
     "start": "node server.js",
@@ -13,8 +14,10 @@
     "bcrypt": "^5.1.1",
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
+    "csv-parser": "^3.0.0",
     "debug": "~2.6.9",
     "dotenv": "^16.4.5",
+    "esm": "^3.2.25",
     "express": "~4.16.4",
     "express-rate-limit": "^7.3.1",
     "express-rate-limiter": "^1.3.1",
@@ -22,27 +25,34 @@
     "fs": "0.0.1-security",
     "http-errors": "~1.6.3",
     "jest": "^29.7.0",
-    "joi": "^17.13.1",
+    "joi": "^17.13.3",
+    "json2csv": "6.0.0-alpha.2",
     "jsonwebtoken": "^9.0.2",
     "mem-fs": "3.0.0",
-    "mongodb": "^6.7.0",
-    "mongoose": "^8.4.0",
+    "mime-types": "^2.1.35",
+    "mongodb": "^6.8.0",
+    "mongoose": "^8.5.1",
+    "mongoose-gridfs": "^1.3.0",
     "morgan": "~1.9.1",
     "node-cache": "^5.1.2",
     "node-fetch": "^3.3.2",
     "node-schedule": "^2.1.1",
-    "nodemon": "^3.1.3",
+    "nodemon": "^3.1.4",
     "papaparse": "^5.4.1",
     "parcel": "^2.12.0",
     "peer": "^1.0.2",
     "pug": "^3.0.3",
-    "svgo": "^3.0.2",
-    "webpack": "^5.92.0",
-    "winston": "^3.13.0"
+    "svgo": "^3.3.2",
+    "tsc": "^2.0.4",
+    "webpack": "^5.93.0",
+    "winston": "^3.13.1"
   },
   "devDependencies": {
-    "@eslint/js": "^9.3.0",
-    "eslint": "^9.3.0",
-    "globals": "^15.3.0"
+    "@eslint/js": "^9.7.0",
+    "@types/fs-extra": "^11.0.4",
+    "@types/mongoose": "^5.11.97",
+    "eslint": "^9.7.0",
+    "globals": "^15.8.0",
+    "typescript": "^5.5.3"
   }
 }
